Fix error caused by nodes with both text and crosslink
When nodes contained both text and a crosslink the concepts created for them did
not get an Xid attribute while they actually are unique concepts and would need
one. I removed the feature that the crosslink text would be adopted into these
concepts and added a check whether the node is an empty crosslink node to prevent
this error. I also added a test case.