Add method get_remote() to xmanager
This method efficiently provides a dictionary that contains all topic ids
and timestamps of all questions and answers in the sheets of an XManager's
document. To realize this method I splitted up taglist into the different
sheets and implemented a new method tag_list() to get all tags from a
document. Furthermore, i moved the function isQuestionNode from utils.py
to xmanager.py and added a new method is_anki_node() that additionally
checks whether the question topic is actually supposed to be in imported
to anki by checking whether it is not an empty node and has answers.