Fix problem with crosslink-answers
Crosslink-Answers were not recognized by get_remote() so I implemented the
function get_answer_nodes() that returns all actual answer_nodes() to a tag
excluding empty nodes and nodes that only contain a crosslink. To check
whether a node only contains a crosslink, I further implemented the
function is_crosslink_node() and moved the functionality to check whether
an xmind link is a crosslink to an independent function is_crosslink(). I
furthermore unfolded the list-comprehension part in get_remote() to
multiple for-loops to improve maintainability.