Add method for processing changes in local questions to sync
This requires to change the question in the ontology so I also implemented
a new method that does this to xontology.py. I changed the getter-methods
in xontology to no more return a list but only the respective data. This
required lots of small indexing changes in different parts of the code.